requirements.txt
requests==2.31.0
selenium==4.15.2
python-dotenv==1.0.0